requirements.txt
absl-py==0.8.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.3.1
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.230
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.0
bert-serving-client==1.8.9
bert-serving-server==1.8.9
bitarray==1.0.1
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
blis==0.2.4
bokeh==1.3.4
boto==2.49.0
boto3==1.9.142
botocore==1.12.220
Bottleneck==1.2.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.3.9
contextlib2==0.6.0
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.0
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
distributed==2.5.1
docutils==0.15.2
editdistance==0.5.3
emmental==0.0.4
en-core-web-sm==2.1.0
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
fasteners==0.15
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
fsspec==0.5.1
future==0.18.2
gast==0.3.2
gevent==1.4.0
glfw==1.8.3
glob2==0.7
gmpy2==2.0.8
google-pasta==0.1.7
GPUtil==1.4.0
greenlet==0.4.15
grpcio==1.24.1
gym==0.15.3
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.23
ipdb==0.12
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mittens==0.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
monotonic==1.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
mujoco-py==2.0.2.7
multipledispatch==0.6.0
multiworld==0.0.0
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.17.4
numpydoc==0.9.1
odo==0.5.1
olefile==0.46
onnx==1.5.0
-e git+https://github.com/onnx/onnx-tensorflow.git@50e453610c9d5ce83c0f78b8a60685caf83bcf35#egg=onnx_tf
openpyxl==3.0.0
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pbr==5.4.3
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.13.0
ply==3.11
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.4.2
pylint==2.4.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
regex==2019.4.14
requests==2.22.0
rope==0.14.0
rsa==3.4.2
ruamel-yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.2
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
snowballstemmer==1.9.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
spacy==2.1.4
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.8
srsly==0.0.5
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==1.14.0
tensorboardX==1.9
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-hub==0.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.0.4
toolz==0.10.0
torch==1.3.1
torchvision==0.3.0
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.7.2
unicodecsv==0.14.1
urllib3==1.24.2
warc==0.2.1
warc3-wet==0.2.3
warcio==1.7.1
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
